requirements.txt
discord.py[voice]==2.4.0
motor==3.4.0
pymongo==4.7.3
pillow==10.4.0
pytesseract==0.3.10
opencv-python==4.10.0.84
numpy==1.26.4
rapidfuzz==3.9.5
python-dotenv==1.0.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1